Add null check for StringIO in case of line execution timeout. (#2318)
# Description

This pull request includes a change to the `write` method in the
`log_manager.py` file. The change adds a check for `None` to avoid
errors when the line execution timeout is reached and all running nodes
are cancelled. This is necessary because for synchronous tools running
in a worker thread, they can't be stopped and the context won't change
in the worker thread as it's a thread-local variable.


# All Promptflow Contribution checklist:
- [X] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [X] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [X] Title of the pull request is clear and informative.
- [X] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

Co-authored-by: robbenwang <robbenwang@microsoft.com>