[internal][bugfix] Create a placeholder for still running line runs (#2200)
# Description

We will ignore those still-running line runs, but when the user runs a
time-consuming flow, he cannot see anything until this line run
terminated - this is very annoyed. So in this PR, we will create a
placeholder for such case, so that user can happily view something, even
before the line(s) terminated.

**Running traces**


![image](https://github.com/microsoft/promptflow/assets/38847871/53b04da0-2351-4cdf-adaf-93a8dbff823a)

**Trace UI url change**

The new url will look like
`http://localhost:<port>/v1.0/ui/traces/?#run=<run>`, there will be a
**#** in the path, which is the way that React/UX route, similar to what
we do with `?` in backend. It's not easy to avoid this breaking change,
and as 1) we can update the print url 2) tracing is still a private
preview feature, so this change shall be acceptable.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.