[SDK] Support override deployment_name and model for python tools (#2229)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request primarily focuses on enhancing the functionality of
the `promptflow` package by introducing the ability to override the
"deployment_name" and "model" in the Python tool. The changes also
include the addition of new test cases to verify the new functionality
and updates to the changelog.

Here are the key changes:

**Bug Fixes and New Features:**

*
[`src/promptflow/CHANGELOG.md`](diffhunk://#diff-41ec3f7c4b5d4c0e670407d3c00a03a6966d7ebf617b1536473e33a12e2bc765R14-R17):
Added a new "Bugs Fixed" section in the changelog to document the new
feature that allows the "deployment_name" to be overridden in the Python
tool.

**Codebase Enhancements:**

*
[`src/promptflow/promptflow/_sdk/_submitter/utils.py`](diffhunk://#diff-38c78274ff4a19381a2129bfd86174d225e0d8537b48c80aa8f6feaca25083abL114-L133):
Modified the `overwrite_connections` function to support the overriding
of "deployment_name" and "model" in the Python tool. The function now
checks for these fields in the connection dictionary and updates the
node inputs accordingly.

**Test Cases:**

*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_run.py`](diffhunk://#diff-94a59a05643476869fa3c6bc45466f1582944a935488075e2e63b6a6a196958fR1362-R1399):
Added new test cases - `test_run_with_deployment_overwrite` and
`test_deployment_overwrite_failure` to validate the new feature of
overriding "deployment_name" and "model".

**Test Configurations:**

*
[`src/promptflow/tests/test_configs/flows/python_tool_deployment_name/flow.dag.yaml`](diffhunk://#diff-f162e17c4af9093d53552b797644d9b1c0b1bf4274910b5f506e4f00caef43acR1-R15):
Added a new test configuration file to support the new test cases.
*
[`src/promptflow/tests/test_configs/flows/python_tool_deployment_name/print_env.py`](diffhunk://#diff-13ed50acd4c00cd4318ccf9d70b3c571b1df072a7e54882c9e84b2124d944ed9R1-R7):
Added a new Python tool `get_result` to be used in the new test cases.
This tool accepts "key", "deployment_name", and "model" as inputs.
# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.