[ProcessPool] Add stderror for process to get error message when process crashed. (#2070)
# Description
**Problem**
When the process crashes unexpectedly, the error information cannot be
obtained.

**Solution**
Redirects stderror to a file, reads the contents of the file, and prints
it when the process crashes.

**Other changes**
Change the exit timing of the spawn process manager in fork mode
_Before_
When all sub-process is not alive, break the spawn process manager.
_After_
When receiving the 'spawned_manager_end' signal from the main process,
exit the spawn process manager.
 _Reason_
When a special input from the user causes the process to crash, we
should give other inputs a chance to try to determine whether the
execution can be successful.
# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.