fix: avoid lazy shutdown for executor proxy in local pfs flow test (#2445)
# Description

This PR is to revert args for `FlowInvoker.__init__`, which change will
bring [compatibility issue with
`mlflow`](https://github.com/mlflow/mlflow/blob/master/mlflow/promptflow/__init__.py#L401).

Given this is a refactor of a private class which has already been
tested explicitly and implicitly in many tests, we haven't added new
tests here.

This pull request primarily focuses on refactoring the flow invocation
process in the `promptflow` package. The changes aim to simplify the
codebase by introducing an `AbstractFlowBase` class, removing redundant
parameters and methods, and adjusting the way flows are initialized and
invoked.

Refactoring flow invocation:

*
[`src/promptflow/promptflow/_sdk/entities/_flow/_flow_context_resolver.py`](diffhunk://#diff-a1a6133992adb765818f3a6474d0f2a1468a609368b4fbc25be1b97e7afa1b14L128-L143):
Removed the `executable_flow` initialization and replaced it with
`resolved_flow` in the `AsyncFlowInvoker` and `FlowInvoker` classes.
Also removed the `flow_path` and `working_dir` parameters from these
classes.

*
[`src/promptflow/promptflow/_sdk/entities/_flow/base.py`](diffhunk://#diff-f2782d63bffa0012d5c70a10a4a3d6c8292ec50bbe3a666e39bf5100a4998c5bL4):
Removed the import of `abc` and added the import of `AbstractFlowBase`.
Changed the `FlowBase` class to inherit from `AbstractFlowBase` and
`SchemaValidatableMixin`, and replaced the initialization of several
instance variables with a call to `AbstractFlowBase.__init__`.
[[1]](diffhunk://#diff-f2782d63bffa0012d5c70a10a4a3d6c8292ec50bbe3a666e39bf5100a4998c5bL4)
[[2]](diffhunk://#diff-f2782d63bffa0012d5c70a10a4a3d6c8292ec50bbe3a666e39bf5100a4998c5bR13)
[[3]](diffhunk://#diff-f2782d63bffa0012d5c70a10a4a3d6c8292ec50bbe3a666e39bf5100a4998c5bL83-L94)

*
[`src/promptflow/promptflow/core/_flow.py`](diffhunk://#diff-a1b7af0df2fd1f9f1d07a45071795e8dc440a303c7847abe5033b9348909712bL10-R19):
Renamed `FlowBase` to `AbstractFlowBase` and added a docstring. Added a
`FlowBase` class that inherits from `AbstractFlowBase`. Removed the
`flow_path` and `working_dir` parameters from the `FlowInvoker` and
`AsyncFlowInvoker` classes.
[[1]](diffhunk://#diff-a1b7af0df2fd1f9f1d07a45071795e8dc440a303c7847abe5033b9348909712bL10-R19)
[[2]](diffhunk://#diff-a1b7af0df2fd1f9f1d07a45071795e8dc440a303c7847abe5033b9348909712bR28-R32)
[[3]](diffhunk://#diff-a1b7af0df2fd1f9f1d07a45071795e8dc440a303c7847abe5033b9348909712bR117-L118)
[[4]](diffhunk://#diff-a1b7af0df2fd1f9f1d07a45071795e8dc440a303c7847abe5033b9348909712bR167-R168)

*
[`src/promptflow/promptflow/core/_serving/app.py`](diffhunk://#diff-642b8b1a75aacde64c3ad2b22c934b38b49e040e60e6d3536077d7c7df1fb19cL16-R20):
Removed the import of `resolve_flow_path` and added the import of
`Flow`. Changed the `FlowInvoker` initialization to load the flow from
`self.project_path` instead of using `self.flow`. Removed the
`flow_path` and `working_dir` parameters from the `FlowInvoker`
initialization.
[[1]](diffhunk://#diff-642b8b1a75aacde64c3ad2b22c934b38b49e040e60e6d3536077d7c7df1fb19cL16-R20)
[[2]](diffhunk://#diff-642b8b1a75aacde64c3ad2b22c934b38b49e040e60e6d3536077d7c7df1fb19cL97-R98)
[[3]](diffhunk://#diff-642b8b1a75aacde64c3ad2b22c934b38b49e040e60e6d3536077d7c7df1fb19cL108-L109)

*
[`src/promptflow/promptflow/core/_serving/flow_invoker.py`](diffhunk://#diff-f01b1126601a80f6099597f2a5820f0722040156377fefd69cd86363d3a62c83L12-R17):
Removed the import of `ExecutableFlow` and added the imports of
`AbstractFlowBase` and `init_executable`. Changed the `flow` parameter
in the `FlowInvoker` class to be of type `Flow` instead of `str` or
`Path`. Replaced the `flow` instance variable with the result of
`init_executable`. Removed the `flow_path` and `working_dir` parameters
from the `FlowInvoker` initialization.
[[1]](diffhunk://#diff-f01b1126601a80f6099597f2a5820f0722040156377fefd69cd86363d3a62c83L12-R17)
[[2]](diffhunk://#diff-f01b1126601a80f6099597f2a5820f0722040156377fefd69cd86363d3a62c83L25-R27)
[[3]](diffhunk://#diff-f01b1126601a80f6099597f2a5820f0722040156377fefd69cd86363d3a62c83L43-R54)
[[4]](diffhunk://#diff-f01b1126601a80f6099597f2a5820f0722040156377fefd69cd86363d3a62c83L68-R68)

*
[`src/promptflow/tests/sdk_cli_test/unittests/test_flow_invoker.py`](diffhunk://#diff-10080d296936c0ec3b532ca6fa629a813c2587d91a9d0b9c7ef987dc83ec9fccR8-R17):
Replaced `init_executable` with `load_flow` to initialize
`EXAMPLE_FLOW`. Removed the `flow_path` and `working_dir` parameters
from the `FlowInvoker` initializations in the test cases.
[[1]](diffhunk://#diff-10080d296936c0ec3b532ca6fa629a813c2587d91a9d0b9c7ef987dc83ec9fccR8-R17)
[[2]](diffhunk://#diff-10080d296936c0ec3b532ca6fa629a813c2587d91a9d0b9c7ef987dc83ec9fccL26-L35)
[[3]](diffhunk://#diff-10080d296936c0ec3b532ca6fa629a813c2587d91a9d0b9c7ef987dc83ec9fccL44-L45)
[[4]](diffhunk://#diff-10080d296936c0ec3b532ca6fa629a813c2587d91a9d0b9c7ef987dc83ec9fccL54-L55)

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.