[Tracing] Modify openai span names (#2459)
Rename openai spans since openai.resources.chat.Completions.create is
too long.

This pull request primarily introduces changes to the tracing
functionality in the promptflow application. The changes allow for more
flexibility and control over the naming of traces, particularly for
asynchronous and synchronous functions. The key changes can be grouped
into two categories: modifications to the function definitions and
adjustments to the function calls.

Modifications to function definitions:


[src/promptflow-tracing/promptflow/tracing/_trace.py](https://github.com/microsoft/promptflow/pull/2424/files#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcL272-R276):
Both the _traced_async and _traced_sync functions were modified to
include an optional name parameter. This allows for custom naming of the
traces. If no name is provided, the function's name is used as the trace
name.
[[1]](https://github.com/microsoft/promptflow/pull/2424/files#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcL272-R276)
[[2]](https://github.com/microsoft/promptflow/pull/2424/files#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcL322-R334)
Adjustments to function calls:


[src/promptflow-tracing/promptflow/tracing/_integrations/_openai_injector.py](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL24-R33):
Several function calls were adjusted to include the new name parameter.
This includes the inject_function_async, inject_function_sync,
inject_async, and inject_sync functions. The name parameter was also
added to the tuples in the _openai_api_list function, which are used to
generate the API and injector.
[[1]](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL24-R33)
[[2]](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL93-R103)
[[3]](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL112-R132)
[[4]](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL141-R146)
[[5]](https://github.com/microsoft/promptflow/pull/2424/files#diff-01b30cb37ac2a91b223014785c6bc7efe007a66efcb97a0dbb36442a94c2054bL179-R182)

---------

Co-authored-by: Heyi <heta@microsoft.com>