[SDK][azure] Update flow.flex.yaml in cloud snapshot (#2749)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request introduces significant changes to the
`promptflow-azure` and `promptflow-devkit` packages, and updates the
`promptflow-recording` test files. The most important changes include
the addition of a method to resolve signatures for flex flows, the
integration of this method into the flow building process, and updates
to the User-Agent and Python version in the test files.

New features and updates:

*
[`src/promptflow-azure/promptflow/azure/_entities/_flow.py`](diffhunk://#diff-61c8df6627e881d340821010be0f85ff5c8080f0f2d40e7527a04e9695aac3f6R134-R141):
Added a new method `_resolve_signature` to resolve signatures for flex
flows. This method uses the `PFClient` to update the signatures.
*
[`src/promptflow-azure/promptflow/azure/_entities/_flow.py`](diffhunk://#diff-61c8df6627e881d340821010be0f85ff5c8080f0f2d40e7527a04e9695aac3f6R163):
Integrated the `_resolve_signature` method into the
`_try_build_local_code` method to update the signatures during the flow
building process.
*
[`src/promptflow-devkit/promptflow/_sdk/operations/_flow_operations.py`](diffhunk://#diff-a6386c44a6e32dbef95fff7e5fcdeb933e5fdea154369d37d7cec10bda0eb226R1238-R1248):
Added a new method `_update_signatures` to update signatures for flex
flows. This method validates the signature and updates the data with the
merged signatures.

Test file updates:

*
[`src/promptflow-recording/recordings/azure/test_run_operations_TestFlowRun_test_eager_flow_cancel.yaml`](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL12-R13):
Updated the User-Agent and Python version in multiple locations
throughout the test file. The User-Agent was updated from
`promptflow-sdk/0.0.1` to `promptflow-azure-sdk/1.8.0.dev0`, and the
Python version was updated from `3.9.18` to `3.11.5`.
[[1]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL12-R13)
[[2]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL56-R57)
[[3]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL108-R109)
[[4]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL162-R163)
[[5]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL201-R203)
[[6]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL285-R286)
[[7]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL339-R340)
[[8]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL506-R508)
[[9]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL469-R475)
[[10]](diffhunk://#diff-e20feb637fd79905d94f9c0181bb3aa42b36031050b23f13c04afb1c6c3ffa5dL551-R555)

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Xingzhi Zhang <37076709+elliotzh@users.noreply.github.com>