Persist Original OpenTelemetry Span Data as JSON in Blob Storage (#2759)
# Description

This PR introduces a change to the way we persist OpenTelemetry span
data. Due to the size limitation of CosmosDB documents, we previously
stored span events in blob storage and other parts in CosmosDB. However,
to provide a complete view of span data for the user, we now persist the
original OpenTelemetry span data as a JSON file in blob storage. This
JSON file is associated with the corresponding CosmosDB record through a
URI.

## Changes Made
- Added a new field `span_json_uri` in the `Span` class to store the URI
of the blob file.
- Implemented a new method `_persist_span_json` in the `Span` class.
This method persists the JSON representation of the Span as a file in
blob storage.
- Updated the `persist` method in the `Span` class to call
`_persist_span_json`, ensuring the original span information is
persisted to storage.

This change enhances the user's ability to request and view full span
data, as the UI can now directly retrieve this information from
CosmosDB.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.