fix: csharp graceful shutdown (#2770)
# Description

In csharp executor service, traces are sent in parallel with exec
response, so it is possible that traces are not fully sent when we try
to destroy the executor proxy. To avoid this, we need to enable graceful
shutdown for csharp.

To enable graceful shutdown (supported in windows only for now), csharp
required a `CTRL_C_EVENT` signal instead of `SIGTERM` signal, which will
forcefully shutdown the csharp service.

On the other hand, directly sending `CTRL_C_EVENT` signal will touch all
processes in target process group so the main process will also be
terminated.

We fix this issue in this PR

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.